+++ source/ProcessManager.Components.Abstractions/EnqueueActorMessages/IEnqueueAcceptedDataDto.cs
@@ -14,7 +14,9 @@
 
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_023_027.V1.Model;
+namespace Energinet.DataHub.ProcessManager.Components.Abstractions.EnqueueActorMessages;
 
-public record CalculationEnqueueActorMessagesV1(Guid CalculationId)
-    : IEnqueueDataDto;
+public interface IEnqueueAcceptedDataDto : IEnqueueDataDto
+{
+    string OriginalActorMessageId { get; init; }
+}