+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,13 @@
 # ProcessManager.Orchestrations.Abstractions Release Notes
 
+## Version 1.18.1
+
+- Use `IEnqueueAcceptedDataDto`, `IEnqueueRejectedDataDto` and `IEnqueueDataDto` marker interfaces.
+
+## Version 1.18.0
+
+- Add custom query `CalculationByIdQueryV1`.
+
 ## Version 1.17.2
 
 - No functional changes.